made number constraint attribute

+++ MADOC.Domain/Ranges/NumberRange.cs
@@ -2,13 +2,13 @@
 {
     public class NumberRange
     {
-        public decimal From { get; set; }
+        public double From { get; set; }
 
-        public decimal To { get; set; }
+        public double To { get; set; }
 
         public NumberRange() { }
 
-        public NumberRange(decimal from, decimal to)
+        public NumberRange(double from, double to)
         {
             this.From = from;
             this.To = to;
